fixed bug in getDayOfWeek
I gave the pointer to my sDate object to getDayOfWeek and so inside the function I manipulated my object directly instead of using copies. Now the already to the object set local variables are used as parameter so it doesnt even matter if they wouldnt be copies in the function since we dont need'em anymore anyway.