Entire exercise 4 push. Testing on Unix needed.
This was nerve-wracking.. a lot of work and for some stuff, I still don't know if it's implemented as he wants it. But it'll definitely be enough to get through. There have to be made some adjustments for unix systems probably since I didn't even try the create file function that works separately for unix. There are countless of comments and TODOs that need to be removed/solved. There are still a lot of improvements to be made and it's really not standardized/consistent throughout the code cuz I was trying a lot of different stuff. I hope I didn't forget to mention anything..