changed waitForEnter func; fixed load bug
Custom useroutput can now be given to waitForEnter func as parameter. We had no real protection when a xml save file had more than MAXAPPOINTMENTS entries. Also when a time/date field was there but with string empty "", memory was unnecessarily allocated, now those fields are checked and if needed an exception is thrown.